Rename and move world related hooks in one place
