+++ src/hooks/worlds/useWorldParams.ts
@@ -7,7 +7,7 @@ export interface WorldEditParams {
   selectedTab?: WorldNavTab;
 }
 
-export const useWorldEditParams = () => {
+export const useWorldParams = () => {
   const {
     worldId,
     selectedTab = WorldNavTab.start,
