+++ src/pages/WorldEditor/WorldEditorStartPanel.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { useWorldById } from "hooks/useWorldById";
+import { useWorldById } from "hooks/worlds/useWorldById";
 
 import { AdminPanel } from "components/organisms/AdminVenueView/components/AdminPanel";
 import { AdminShowcase } from "components/organisms/AdminVenueView/components/AdminShowcase";
